Move VecInput into conn::receive
