+++ rustls/src/msgs/deframer/mod.rs
@@ -10,8 +10,8 @@ use crate::msgs::{HEADER_SIZE, read_opaque_message_header};
 
 mod buffers;
 use buffers::Coalescer;
+pub use buffers::SliceInput;
 pub(crate) use buffers::{Delocator, Locator};
-pub use buffers::{SliceInput, VecInput};
 
 pub fn fuzz_deframer(data: &[u8]) {
     let mut buf = data.to_vec();
